requirements.txt
mysql_connector_repackaged==0.3.1
numpy==1.24.3
pandas==2.0.1
plotly==5.14.1
PyMySQL==1.0.3
Requests==2.31.0
streamlit==1.23.1
streamlit_lottie==0.0.5
streamlit_option_menu==0.3.6
